requirements.txt
kafka-python-ng
python-dotenv==1.0.0